+++ cybsi/api/observable/aggregate_section.py
@@ -48,6 +48,8 @@ def _convert_attribute_value_type(
         AttributeNames.Campaigns: DictionaryItemCommonView,
         AttributeNames.ThreatActors: DictionaryItemCommonView,
         AttributeNames.AffectedCountries: DictionaryItemCommonView,
+        AttributeNames.ExploitedVulnerabilities: DictionaryItemCommonView,
+        AttributeNames.TargetedSectors: DictionaryItemCommonView,
     }
 
     return _attribute_value_types[attribute_name](val)