+++ cybsi/api/observable/entities_api.py
@@ -55,6 +55,8 @@ _attr_value_kebab_converters: Dict[AttributeNames, str] = {
     AttributeNames.Campaigns: "campaigns",
     AttributeNames.ThreatActors: "threat-actors",
     AttributeNames.AffectedCountries: "affected-countries",
+    AttributeNames.ExploitedVulnerabilities: "exploited-vulnerabilities",
+    AttributeNames.TargetedSectors: "targeted-sectors",
 }